requirements.txt
streamlit
openai
pymupdf
python-docx
pandas
